Updated the data model classes

+++ PlaceSearcher/app/src/main/java/com/tulikab/placesearcher/data/VenueLocation.java
@@ -4,8 +4,8 @@ public class VenueLocation {
 
     private String address;
     private String crossStreet;
-    private long lat;
-    private long lng;
+    private double lat;
+    private double lng;
     private int postalCode;
     private String countryCode;
     private String city;
@@ -29,19 +29,19 @@ public class VenueLocation {
         this.crossStreet = crossStreet;
     }
 
-    public long getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(long lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public long getLng() {
+    public double getLng() {
         return lng;
     }
 
-    public void setLng(long lng) {
+    public void setLng(double lng) {
         this.lng = lng;
     }
